Factor never service out of test_connection to fixup the dependency issues
