+++ rust-runtime/smithy-client/src/test_connection.rs
@@ -7,10 +7,9 @@
 // TODO
 #![allow(missing_docs)]
 
-use crate::test_connection::stream::EmptyStream;
 use http::header::{HeaderName, CONTENT_TYPE};
 use http::{Request, Uri};
-use hyper::client::HttpConnector;
+
 use protocol_test_helpers::{assert_ok, validate_body, MediaType};
 use smithy_async::future::never::Never;
 use smithy_http::body::SdkBody;
@@ -46,125 +45,8 @@ impl CaptureRequestReceiver {
     }
 }
 
-/// A service that will never return whatever it is you want
-///
-/// Returned futures will return Pending forever
-#[non_exhaustive]
-#[derive(Debug)]
-pub struct NeverService<R> {
-    _resp: PhantomData<R>,
-}
-
-impl<R> Clone for NeverService<R> {
-    fn clone(&self) -> Self {
-        Self {
-            _resp: Default::default(),
-        }
-    }
-}
-
-impl<R> NeverService<R> {
-    pub fn new() -> Self {
-        NeverService {
-            _resp: Default::default(),
-        }
-    }
-}
-
-pub mod stream {
-    use hyper::client::connect::{Connected, Connection};
-    use std::io::Error;
-    use std::pin::Pin;
-    use std::sync::atomic::AtomicBool;
-    use std::task::{Context, Poll};
-    use tokio::io::{AsyncRead, AsyncWrite, ReadBuf};
-
-    /// A stream that will never return or accept any data
-    #[non_exhaustive]
-    pub struct EmptyStream;
-
-    impl EmptyStream {
-        pub fn new() -> Self {
-            Self
-        }
-    }
-
-    impl AsyncRead for EmptyStream {
-        fn poll_read(
-            self: Pin<&mut Self>,
-            _cx: &mut Context<'_>,
-            _buf: &mut ReadBuf<'_>,
-        ) -> Poll<std::io::Result<()>> {
-            Poll::Pending
-        }
-    }
-
-    impl AsyncWrite for EmptyStream {
-        fn poll_write(
-            self: Pin<&mut Self>,
-            _cx: &mut Context<'_>,
-            _buf: &[u8],
-        ) -> Poll<Result<usize, Error>> {
-            Poll::Pending
-        }
-
-        fn poll_flush(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Result<(), Error>> {
-            Poll::Pending
-        }
-
-        fn poll_shutdown(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Result<(), Error>> {
-            Poll::Pending
-        }
-    }
-
-    impl Connection for EmptyStream {
-        fn connected(&self) -> Connected {
-            Connected::new()
-        }
-    }
-}
-
-pub type NeverConnected = NeverService<TcpStream>;
-
-/// A service that will connect but never send any data
-#[derive(Clone)]
-pub struct NeverReplies;
-impl NeverReplies {
-    pub fn new() -> Self {
-        Self
-    }
-}
-
-impl tower::Service<Uri> for NeverReplies {
-    type Response = stream::EmptyStream;
-    type Error = BoxError;
-    type Future = std::future::Ready<Result<Self::Response, Self::Error>>;
-
-    fn poll_ready(&mut self, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
-        Poll::Ready(Ok(()))
-    }
-
-    fn call(&mut self, _req: Uri) -> Self::Future {
-        std::future::ready(Ok(EmptyStream::new()))
-    }
-}
-
-impl<Req, Resp> tower::Service<Req> for NeverService<Resp> {
-    type Response = Resp;
-    type Error = BoxError;
-    type Future = Pin<Box<dyn Future<Output = Result<Self::Response, Self::Error>> + Send>>;
-
-    fn poll_ready(&mut self, _cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
-        Poll::Ready(Ok(()))
-    }
-
-    fn call(&mut self, _req: Req) -> Self::Future {
-        Box::pin(async move {
-            Never::new().await;
-            unreachable!()
-        })
-    }
-}
+#[doc(inline)]
+pub use crate::never;
 
 impl tower::Service<http::Request<SdkBody>> for CaptureRequestHandler {
     type Response = http::Response<SdkBody>;
@@ -374,7 +256,8 @@ where
 
 #[cfg(test)]
 mod tests {
-    use crate::test_connection::{capture_request, NeverService, TestConnection};
+    use crate::never::NeverService;
+    use crate::test_connection::{capture_request, never::NeverService, TestConnection};
     use crate::{BoxError, Client};
     use hyper::service::Service;
     use smithy_http::body::SdkBody;
