+++ rust-runtime/smithy-client/src/timeout.rs
@@ -2,18 +2,19 @@
 //!
 //! While timeout configuration is unstable, this module is in smithy-client.
 //!
-//! As timeout and HTTP configuration stabilizes, this will move to aws-types and become
+//! As timeout and HTTP configuration stabilizes, this will move to aws-types and become a part of
+//! HttpSettings.
 use std::time::Duration;
 
 /// Timeout Configuration
 #[derive(Default, Debug, Clone)]
 #[non_exhaustive]
-pub struct Config {
+pub struct Settings {
     connect_timeout: Option<Duration>,
     http_read_timeout: Option<Duration>,
 }
 
-impl Config {
+impl Settings {
     /// Create a new timeout configuration
     ///
     /// The current default (subject to change) is no timeouts
